refactor and restructure db
